Create sub array
Check if the sub array with given product exists int he array. Given an array of both positive and negative integers and a number k. The task is to check if any subarray with produck k is present in the array or not.
EXAMPLE :
INPUT : arr[]={-2, -1, 3, -4, -5}, k=2
OUTPUT : YES

INPUT : arr[]={3, -1, -1, -1, 5}, k=3
OUTPUT : YES